Add instruction for running superdoku.sh manually on Windows.
Recommended next steps are to implement a timeout on the client side
when a request is in progress so that the user doesn't have to stare at
an infinite loading screen when there is a server issue.

Also along the same lines, should implement the ability to go back or
cancel a request while an image is being scanned in the
ImageProcessorScene.