Refactor solve function to be asynchronous
Make use of a Dispatcher to run UnityEngine functions on the main thread
from an async thread.

LESSON LEARNED: Make sure the Update function of a MonoBehavior class is
not contradicting/conflicting with an asynchronous function (e.g.,
SolverPuzzle in PuzzleGrid class); solution in this case was to use a
boolean flag to turn off the update function only when the puzzle was
being solved.

Next steps are to fix minor bug with the update function not getting
turned off in time before a few digits get displayed to the puzzle
(doesn't effect solution).

Resources:
https://answers.unity.com/questions/305882/how-do-i-invoke-functions-on-the-main-thread.html
(Dispatcher)

https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/concepts/async/
(Understanding asynchronous thread implementations using C#)